Packages renamed and flyway dependancy + flyway bundle added
