+++ pom.xml
@@ -20,11 +20,12 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <dropwizard.version>1.3.3</dropwizard.version>
+        <dropwizard-flyway.version>0.9.2-3</dropwizard-flyway.version>
         <jooq.version>3.11.0</jooq.version>
         <sqlite.version>3.23.1</sqlite.version>
         <swagger.version>1.3.1-1</swagger.version>
         <jooq-dropwizzard-bundle.version>1.2.0-0</jooq-dropwizzard-bundle.version>
-        <mainClass>com.arenteria.test.TruelogicTestApplication</mainClass>
+        <mainClass>com.arenteria.test.truelogic.TruelogicTestApplication</mainClass>
     </properties>
 
     <dependencyManagement>
@@ -52,6 +53,17 @@
             <groupId>com.bendb.dropwizard</groupId>
             <artifactId>dropwizard-jooq</artifactId>
             <version>${jooq-dropwizzard-bundle.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.dropwizard.modules</groupId>
+            <artifactId>dropwizard-flyway</artifactId>
+            <version>${dropwizard-flyway.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>io.dropwizard</groupId>
+                    <artifactId>dropwizard-db</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
         <dependency>
             <groupId>org.jooq</groupId>
@@ -191,7 +203,7 @@
                             <excludes></excludes>
                         </database>
                         <target>
-                            <packageName>org.jooq.sources</packageName>
+                            <packageName>com.arenteria.test.truelogic.db</packageName>
                             <directory>target/generated-sources/jooq</directory>
                         </target>
                         <generate>
