--- src/main/java/com/arenteria/test/core/constant/BookshelfConstants.java
@@ -1,7 +1,0 @@
-package com.arenteria.test.core.constant;
-
-
-public class BookshelfConstants {
-    public static final String ERROR_BOOK_NOT_FOUND = "ERROR_BOOK_NOT_FOUND";
-}
-
