--- src/main/java/com/arenteria/test/core/exception/DaoException.java
@@ -1,4 +1,0 @@
-package com.arenteria.test.core.exception;
-
-public class DaoException extends Exception{
-}
