--- src/main/java/com/arenteria/test/core/exception/ServiceException.java
@@ -1,4 +1,0 @@
-package com.arenteria.test.core.exception;
-
-public class ServiceException extends Exception {
-}
