--- src/main/java/com/arenteria/test/service/BookshelfService.java
@@ -1,18 +1,0 @@
-package com.arenteria.test.service;
-
-import com.arenteria.test.core.exception.ServiceException;
-import com.arenteria.test.domain.dto.BookDTO;
-import java.util.Collection;
-
-public interface BookshelfService {
-
-    Collection<BookDTO> findBookByTittle(String tittle);
-
-    Collection<BookDTO> findBookByDescription(String description);
-
-    Collection<BookDTO> findAllBooks();
-
-    BookDTO findBookById(Integer id);
-
-    void saveBook(BookDTO bookDTO) throws ServiceException;
-}
