+++ src/main/java/com/arenteria/test/truelogic/core/constant/BookshelfConstants.java
@@ -1,0 +1,7 @@
+package com.arenteria.test.truelogic.core.constant;
+
+
+public class BookshelfConstants {
+    public static final String ERROR_BOOK_NOT_FOUND = "ERROR_BOOK_NOT_FOUND";
+}
+
