+++ src/main/java/com/arenteria/test/truelogic/core/exception/InternalServerErrorException.java
@@ -1,0 +1,11 @@
+package com.arenteria.test.truelogic.core.exception;
+
+public class InternalServerErrorException extends RestException {
+
+    private int code;
+
+    public InternalServerErrorException (int code, String msg) {
+        super(code, msg);
+        this.code = code;
+    }
+}
