+++ src/main/java/com/arenteria/test/truelogic/core/exception/NotFoundException.java
@@ -1,0 +1,11 @@
+package com.arenteria.test.truelogic.core.exception;
+
+public class NotFoundException extends RestException {
+
+    private int code;
+
+    public NotFoundException (int code, String msg) {
+        super(code, msg);
+        this.code = code;
+    }
+}
