+++ src/main/java/com/arenteria/test/truelogic/core/exception/ServiceException.java
@@ -1,0 +1,4 @@
+package com.arenteria.test.truelogic.core.exception;
+
+public class ServiceException extends Exception {
+}
